ged/requirements.txt
asgiref==3.9.2
Django==5.2.6
django-cors-headers==4.9.0
djangorestframework==3.16.1
fpdf==1.7.2
psycopg2==2.9.10
sqlparse==0.5.3
tzdata==2025.2